+++ crates/intrinsic-test/src/arm/format.rs
