+++ crates/intrinsic-test/src/common/types.rs
@@ -1,0 +1,5 @@
+#[derive(Debug, PartialEq)]
+pub enum Language {
+    Rust,
+    C,
+}
